requirements.txt
streamlit==1.26.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
